More tidying.
